Move booksform and bookslist to containers folder
